+++ rex/utils/logging.py
@@ -29,7 +29,7 @@ class MultiProcessLogger(_Logger):
 
 
 # logger depth=1 to show original filename and call info under logger inheritance
-logger = MultiProcessLogger(_Core(), None, 1, False, False, True, False, True, None, {})
+logger = MultiProcessLogger(_Core(), None, 1, False, False, True, False, True, [], {})
 logger.add(
     lambda msg: _tqdm.write(msg, end=""), colorize=True, backtrace=True, diagnose=True
 )