+++ rex/data/dataset.py
@@ -2,6 +2,8 @@ from typing import Iterable, Iterator, Optional
 
 from torch.utils.data import Dataset, IterableDataset
 
+from rex.utils.logging import logger
+
 
 class CachedDataset(Dataset):
     def __init__(self, data: Iterable) -> None:
@@ -34,13 +36,18 @@ class CachedBagREDataset(Dataset):
 
 class StreamTransformDataset(Dataset):
     def __init__(
-        self, data: Iterable, transform, debug: Optional[bool] = False
+        self,
+        data: Iterable,
+        transform,
+        dataset_name: str = None,
+        debug: Optional[bool] = False,
     ) -> None:
         super().__init__()
         if debug:
             data = data[:128]
         self.data = data
         self.transform = transform
+        self.dataset_name = dataset_name
 
     def __getitem__(self, index: int):
         return self.transform(self.data[index])
@@ -51,18 +58,36 @@ class StreamTransformDataset(Dataset):
 
 class StreamReadDataset(IterableDataset):
     def __init__(
-        self, data_iterator: Iterator, transform, debug: Optional[bool] = False
+        self,
+        data_iterator: Iterator,
+        transform,
+        dataset_name: str = None,
+        debug: Optional[bool] = False,
+        num_samples: int = 3,
     ) -> None:
         super().__init__()
 
+        self.dataset_name = dataset_name
         self.transform = transform
         self.data_iterator = data_iterator
         self.debug = debug
         self.cnt = 0
+        self.num_samples = num_samples
 
     def __iter__(self):
         for item in self.data_iterator:
             if self.debug and self.cnt >= 500:
                 raise StopIteration
-            yield self.transform(item)
-            self.cnt += 1
+            results = self.transform(item, dataset_name=self.dataset_name)
+
+            if self.cnt < self.num_samples:
+                logger.opt(colors=False).debug(f"{results}")
+
+            if isinstance(results, dict):
+                self.cnt += 1
+                yield results
+            elif isinstance(results, list):
+                self.cnt += len(results)
+                yield from results
+            else:
+                continue